requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
scikit-learn==1.6.1
xgboost==2.0.3
pandas==2.2.0
numpy==1.26.3
google-generativeai==0.3.0
joblib==1.3.2
streamlit==1.32.0
python-multipart==0.0.9
python-dotenv==1.0.0
groq